move Decode like messages back to helpers package ref #62
Conflicts:
	src/device-wallet/messages.go
